Separate SerialMonitor implementation module
